+TotalExpensesTest
Added a total_expenses test case.
Interestingly, didn't need to re-inititialise Analyzer in this method, as it was already made in the previous test function.
I don't know if this is a quirk of Python or Unittests, but I tested it in my own debugger and in the Command prompt and it seemed fine so I'll take it