testAnalyzer implementation
So turns out the analyzer I was referring to before in the test_total_expenses case was referring to the analyzer from main, not the one from the test case.

So I've renamed the test case analyser to testAnalyzer (for clarity's sake) and made a local testAnalyzer for each test function.

Also implemented a test_profit functions